crystalbet/backend/app/requirements.txt
fastapi==0.95.2
pydantic==1.10.7
uvicorn[standard]==0.22.0
pymongo==4.5.0
dnspython==2.3.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.1.0
starlette==0.20.4
httpx==0.24.1
loguru==0.7.0

# Optional
motor==3.1.0
authlib==1.2.1
pytest==7.4.2
pytest-asyncio==0.20.3
stripe==4.5.0
